Move cell interface to dedicated submodule
