+++ docs/api/autodoc/spglib.cell.rst
@@ -1,0 +1,8 @@
+spglib.cell module
+==================
+
+.. automodule:: spglib.cell
+   :members:
+   :private-members:
+   :show-inheritance:
+   :undoc-members:
