+++ python/spglib/__init__.py
@@ -7,6 +7,7 @@
 #  backwards compatibility
 from ._internal import *  # noqa: F403
 from ._version import __version__, __version_tuple__  # noqa: F401
+from .cell import *  # noqa: F403
 from .kpoints import *  # noqa: F403
 from .msg import *  # noqa: F403
 from .reduce import *  # noqa: F403
@@ -14,14 +15,11 @@ from .spg import *  # noqa: F403
 
 # fmt: off
 from .spglib import (  # noqa: F401
-    find_primitive,
     get_error_message,
     get_hall_number_from_symmetry,
     get_version,
-    refine_cell,
     spg_get_commit,
     spg_get_version,
     spg_get_version_full,
-    standardize_cell,
 )
 # fmt: on
